Fix linter errors and enhance codebase type safety
- Resolve AuthException import issues and remove unused variables in authentication service
- Fix num to double conversion in performance service
- Add null safety checks in restaurant card widget
- Update place order usecase with proper CartItemEntity types
- Fix const constructor issues in food details and search screens
- Generate JSON serialization files for order and restaurant detail models
- Add new services: Agora, Geocoding, Payment, Receipt, WebRTC
- Implement new features: Category browse, Notifications, Settings, Splash screen enhancements
- Add comprehensive test coverage for new services
- Update pubspec dependencies and lock file
- Add detailed documentation files for analysis, implementation, and integrations

+++ food_delivery_app/lib/core/services/performance_service.dart
@@ -131,10 +131,10 @@ class PerformanceService {
       min: min.toInt(),
       max: max.toInt(),
       mean: mean,
-      median: median,
-      p90: p90,
-      p95: p95,
-      p99: p99,
+      median: median.toDouble(),
+      p90: p90.toDouble(),
+      p95: p95.toDouble(),
+      p99: p99.toDouble(),
     );
   }
